// examples/type.ts
export enum DayCount {
    ACT_ACT_ISDA = "Actual/Actual ISDA",
    ACT_365F = "Actual/365F",
    ACT_365_25 = "Actual/365.25",
    ACT_364 = "Actual/364",
    ACT_360 = "Actual/360",
    THIRTY_360_ISDA = "30/360 ISDA",
    THIRTY_E_360 = "30E/360",
    THIRTY_E_PLUS_360 = "30E+/360",
    THIRTY_E_360_ISDA = "30E/360 ISDA",
    THIRTY_U_360 = "30U/360",
    NL_365 = "NL/365",
    NL_360 = "NL/360",
}